fix(tests): Integrated ts tests into test runner, moved location.
