--- runtime/src/tests/functionalTests/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es6",
-    "sourceMap": true,
-    "allowJs": true,
-    "checkJs": true,
-    "strict": true,
-    "strictNullChecks": true
-  },
-  "exclude": [
-    "src/tests/",
-    "./*.js"
-  ]
-}
