--- src/tests/constants.ts
@@ -1,8 +1,0 @@
-/*
- * Constants Definition
- */
-export class Constants {
-    public WAIT: number = 30000;
-    public LONG_WAIT: number = 60000;
-}
-
