+++ src/tests/specs/smokeTest.spec.ts
