+++ src/tests/ui/planner/sidepanel.ts
